requirements.txt
Eve==0.6.1
Flask-PyMongo~=0.4
requests~=2.8
gunicorn~=19.4